chore: remove roslyn plugin and switch to built-in roslyn_ls LSP
- Remove roslyn-nvim from nix plugins
- Remove custom roslyn.lua setup configuration
- Add roslyn_ls to LSP servers configuration
- Remove x86_64-darwin from supported systems
