+++ nix/plugins.nix
@@ -60,7 +60,6 @@ with pkgs.vimPlugins;
   friendly-snippets
   ## Language specific
   nvim-jdtls
-  roslyn-nvim
 
   # lualine.lua
   lualine-nvim
