+++ nvim/lua/config/plugins/lsp/servers.lua
@@ -152,6 +152,9 @@ local servers = {
       },
     },
   },
+
+  --- https://github.com/neovim/nvim-lspconfig/blob/master/lsp/roslyn_ls.lua
+  roslyn_ls = {},
 }
 
 return servers
