+++ nvim/tests/integration/lsp_dotnet_roslyn.lua
@@ -8,14 +8,14 @@ local sdk_major = assert(vim.env.INTEGRATION_TEST_DOTNET_SDK_MAJOR)
 runner.define({
   describe = string.format("Roslyn .NET %s integration test", sdk_major),
   it = "attaches and clears stale Helper diagnostics after external file creation",
-  client_name = "roslyn",
+  client_name = "roslyn_ls",
   timeout_ms = 60000,
   request_timeout_ms = 5000,
   probes = {
     {
       type = "attach",
       assert = function(client)
-        assert.are.equal("roslyn", client.name)
+        assert.are.equal("roslyn_ls", client.name)
       end,
     },
     {
